get_position/requirements.txt
is-msgs==1.1.10
is-wire==1.2.0
vine==1.3.0
Pillow
torch


